% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noegletal_get.R
\name{noegletal_get}
\alias{noegletal_get}
\title{Automatically get requested data from noegletal.dk}
\usage{
noegletal_get(muni_codes, years, variable_ids)
}
\arguments{
\item{muni_codes}{A vector of municipalities to include in the dataset, as identified by their official municipality codes.}

\item{years}{A vector of years to include in the dataset (2007-current).}

\item{variables}{A vector of variables (nøgletal) to include in the dataset, as identified by the internal noegletal.dk variable id's.}
}
\value{
A tidy \code{tibble} with one row for each municipality-year, and one column for each included variable (nøgletal).
}
\description{
\code{noegletal_get} make it possible to get data from noegletal.dk directly from the comfort of R.
}
