% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noegletalR-package.R
\docType{package}
\name{noegletalR-package}
\alias{noegletalR}
\alias{noegletalR-package}
\title{noegletalR: Tidy Tibbles of Noegletal}
\description{
A simple R utility that parses csv files from noegletal.dk into tidy tibbles.
}
\author{
\strong{Maintainer}: Frederik Larsen \email{frlars21@student.aau.dk}

}
\keyword{internal}
