% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noegletal_get.R
\name{noegletal_get}
\alias{noegletal_get}
\title{Get data from noegletal.dk}
\usage{
noegletal_get(
  muni_codes = ALLOWED_MUNI_CODES,
  years = ALLOWED_YEARS,
  variable_ids
)
}
\arguments{
\item{muni_codes}{Vector of municipality codes}

\item{years}{Vector of years}

\item{variable_ids}{Vector of variable IDs}
}
\value{
Tidy tibble with requested data
}
\description{
Get data from noegletal.dk
}
